bug: type 해결 /find 오류
