+++ frontend/package.json
@@ -10,15 +10,15 @@
   },
   "dependencies": {
     "@heroicons/react": "^2.2.0",
-    "next": "15.1.7",
+    "next": "^15.2.3",
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
-    "@types/node": "20.17.19",
-    "@types/react": "19.0.10",
-    "@types/react-dom": "^19",
+    "@types/node": "^20.17.24",
+    "@types/react": "^19.0.12",
+    "@types/react-dom": "^19.0.4",
     "autoprefixer": "^10.4.21",
     "eslint": "^9",
     "eslint-config-next": "15.1.7",
